# Fix multinomial CUDA misalignment and non-deterministic behavior