--- a/aten/src/ATen/native/cuda/MultinomialKernel.cu
+++ b/aten/src/ATen/native/cuda/MultinomialKernel.cu
@@ -182,6 +182,12 @@ __global__ void sampleMultinomialOnce(
   accscalar_t accZero = static_cast<accscalar_t>(0);
   scalar_t zero = static_cast<scalar_t>(0);
 
+  // Valid threads are those that actually process an element
+  bool is_valid_thread = threadIdx.x < categories;
+  // Block size without padding threads
+  // (so that its size is a multiple of C10_WARP_SIZE)
+  int valid_block_size = (categories < blockDim.x) ? categories : blockDim.x;
+
   for (int64_t curDist = blockIdx.x;
        curDist < distributions; curDist += gridDim.x) {
     // Each block handles one distribution
@@ -236,46 +242,50 @@ __global__ void sampleMultinomialOnce(
                              static_cast<accscalar_t>(dist[curDist * stride_dist + cat * stride_categories]) / sum :
                              accZero;
 
-      smem[threadIdx.x] = dist_val;
+      if (is_valid_thread) {
+        smem[threadIdx.x] = dist_val;
+      }
       __syncthreads();
 
       // Perform an inclusive prefix sum of the shared memory contents
-      for (int offset = 1; offset < blockDim.x; offset *= 2) {
+      for (int offset = 1; offset < valid_block_size; offset *= 2) {
         accscalar_t val = accZero;
 
-        if (threadIdx.x >= offset) {
+        if (is_valid_thread && threadIdx.x >= offset) {
           val = smem[threadIdx.x - offset] + smem[threadIdx.x];
         }
 
         __syncthreads();
-        if (threadIdx.x >= offset) {
+        if (is_valid_thread && threadIdx.x >= offset) {
           smem[threadIdx.x] = val;
         }
         __syncthreads();
       }
 
-      // Each thread will check to see if the sample falls in its
-      // bucket
-      scalar_t curBucket =
-          static_cast<scalar_t>(smem[threadIdx.x] + prevHighProb);
-      scalar_t prevBucket = static_cast<scalar_t>(
-          threadIdx.x == 0 ? prevHighProb
-                           : smem[threadIdx.x - 1] + prevHighProb);
-      bool inBucket =
-          (cat < categories) &&
-          (!(sample >= curBucket) &&
-           (sample >= prevBucket) &&
-           (dist_val > zero));
+      if (is_valid_thread) {
+        // Each thread will check to see if the sample falls in its
+        // bucket
+        scalar_t curBucket =
+            static_cast<scalar_t>(smem[threadIdx.x] + prevHighProb);
+        scalar_t prevBucket = static_cast<scalar_t>(
+            threadIdx.x == 0 ? prevHighProb
+                            : smem[threadIdx.x - 1] + prevHighProb);
+        bool inBucket =
+            (cat < categories) &&
+            (!(sample >= curBucket) &&
+            (sample >= prevBucket) &&
+            (dist_val > zero));
 
-      if (inBucket) {
-        // We're done; we have the sample
-        // Torch indices are 1-based
-        atomicMax(&foundPos, cat);
-        found = true;
+        if (inBucket) {
+            // We're done; we have the sample
+            // Torch indices are 1-based
+            atomicMax(&foundPos, cat);
+            found = true;
+        }
+
+        // Store the previous scan's high value for future use
+        prevHighProb = prevHighProb + smem[valid_block_size - 1];
       }
-
-      // Store the previous scan's high value for future use
-      prevHighProb = prevHighProb + smem[blockDim.x - 1];
 
       __syncthreads();
     }
@@ -327,8 +337,7 @@ void multinomial_with_replacement_kernel_impl(
     int maxThreads = props->maxThreadsPerBlock;
     int maxShared = props->sharedMemPerBlock;
     int requiredShared =
-        (numCategories < maxThreads ? min(2, numCategories) : maxThreads) *
-        sizeof(accscalar_t);
+        std::min(maxThreads, std::max(2, numCategories)) * sizeof(accscalar_t);
 
     if (n_sample == 1 && maxShared >= requiredShared) {
       // Optimized allocation-free implementation
@@ -340,8 +349,9 @@ void multinomial_with_replacement_kernel_impl(
                                           self_v.options().pinned_memory_opt());
       at::native::uniform_(sampled, 0.0, 1.0, generator);
 
-      dim3 block(numCategories < maxThreads ? numCategories : maxThreads);
-      dim3 grid(numDist < numSM * 4 ? numDist : numSM * 4);
+      int warps = at::cuda::ATenCeilDiv(numCategories, C10_WARP_SIZE);
+      dim3 block(std::min(warps * C10_WARP_SIZE, maxThreads));
+      dim3 grid(std::min(static_cast<int>(numDist), numSM * 4));
 
       sampleMultinomialOnce<scalar_t, accscalar_t>
           <<<grid, block,