@ysiraichi 2021-04-06
Here are a few relevant benchmarks.

- Each row was executed `1000` times
- The "Time" columns represent the accumulated execution time

<details>
<summary>Benchmark Script</summary>

```python
import timeit

dtypes = [
    'float16',
    'float32',
    'float64',
]

shapes = [
    (1000, 1000),
    (1000, 2000),
    (1000, 4000),
    (1000, 8000),
    (1000, 16000),
]

device = 'cuda'
number = 1000

def run(shape, dtype, printit=True):
    print(timeit.timeit(
        setup=f"import torch; a = torch.rand(*{shape}, dtype=torch.{dtype}, device='{device}');",
        stmt=f"torch.multinomial(a, num_samples=1, replacement=True)",
        number=number
    ))

for shape in shapes:
    for dtype in dtypes:
        run(shape, dtype)

```
</details>


| shape         | dtype   | [base] Time (s) | [fix] Time (s) |
|---------------|---------|:---------------:|:--------------:|
| (1000, 1000)  | float16 |           0.043 |          0.040 |
| (1000, 1000)  | float32 |           0.057 |          0.062 |
| (1000, 1000)  | float64 |           0.131 |          0.142 |
| (1000, 2000)  | float16 |           0.126 |          0.124 |
| (1000, 2000)  | float32 |           0.085 |          0.083 |
| (1000, 2000)  | float64 |           0.171 |          0.169 |
| (1000, 4000)  | float16 |           0.199 |          0.197 |
| (1000, 4000)  | float32 |           0.139 |          0.136 |
| (1000, 4000)  | float64 |           0.271 |          0.267 |
| (1000, 8000)  | float16 |           0.329 |          0.323 |
| (1000, 8000)  | float32 |           0.237 |          0.231 |
| (1000, 8000)  | float64 |           0.478 |          0.470 |
| (1000, 16000) | float16 |           0.583 |          0.574 |
| (1000, 16000) | float32 |           0.438 |          0.428 |
| (1000, 16000) | float64 |           0.888 |          0.875 |