@kshitij12345 2021-04-07
Just a passing comment.

I think there is an issue with the benchmarking script. The CUDA kernel launches are async, thus the script is just checking for launch overhead but not the actual time taken by the operator.

I think it would be better to use pytorch's benchmarking utility (which takes that into account)

Quote from the docs
> The difference in the results between timeit and benchmark modules is because the timeit module is not synchronizing CUDA and is thus only timing the time to launch the kernel. PyTorch’s benchmark module does the synchronization for us.

https://pytorch.org/tutorials/recipes/recipes/benchmark.html#benchmarking-with-torch-utils-benchmark-timer