@ysiraichi 2021-04-07
Here are a few relevant benchmarks. Each row was repeatedly ran until completing, at least, 1s. 

<details>
<summary>Benchmark Script</summary>

```python
import torch.utils.benchmark as benchmark
import torch

def run(a):
    torch.multinomial(a, num_samples=1, replacement=True)

dtypes = [
    torch.float16,
    torch.float32,
    torch.float64,
]

shapes = [
    (1000, 1000),
    (1000, 2000),
    (1000, 4000),
    (1000, 8000),
    (1000, 16000),
]

device = 'cuda'
number = 1000

for shape in shapes:
    for dtype in dtypes:
        a = torch.rand(*shape, dtype=dtype, device=device)
        m = benchmark.Timer(
            setup=f"from __main__ import run",
            stmt=f"run(a)",
            globals={"a": a},
        ).blocked_autorange(min_run_time=1)
        print(f"| {shape} | {dtype} | {m.mean * 1000:.3f} |")
```
</details>


| shape         | dtype   | [base] Time (ms) | [fix] Time (ms) |
|---------------|---------|:----------------:|:---------------:|
| (1000, 1000)  | float16 |            0.057 |           0.056 |
| (1000, 1000)  | float32 |            0.056 |           0.055 |
| (1000, 1000)  | float64 |            0.170 |           0.167 |
| (1000, 2000)  | float16 |            0.084 |           0.082 |
| (1000, 2000)  | float32 |            0.088 |           0.086 |
| (1000, 2000)  | float64 |            0.264 |           0.257 |
| (1000, 4000)  | float16 |            0.140 |           0.136 |
| (1000, 4000)  | float32 |            0.139 |           0.135 |
| (1000, 4000)  | float64 |            0.423 |           0.416 |
| (1000, 8000)  | float16 |            0.240 |           0.233 |
| (1000, 8000)  | float32 |            0.241 |           0.234 |
| (1000, 8000)  | float64 |            0.739 |           0.725 |
| (1000, 16000) | float16 |            0.439 |           0.425 |
| (1000, 16000) | float32 |            0.450 |           0.438 |
| (1000, 16000) | float64 |            1.371 |           1.346 |