@kurtamohler 2021-04-09
@ysiraichi, could you add a test that exercises the CUDA misalignment error that you fixed? Probably just the example from the issue description would be good:
```python
torch.multinomial(torch.rand(2, 3, dtype=torch.float16, device="cuda"), num_samples=1, replacement=True)
```

Just something to think about, you could consider adding a test for determinism as well. A caveat though is that a test cannot actually prove determinism. Also, if it typically takes a long runtime to actually see the current non-deterministic behavior, then it may not be worth adding a test.